requirements.txt
requests
pyjwt
pydantic
